backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8
cloudinary==1.36.0
python-socketio==5.10.0
redis==5.0.1
celery==5.3.4
pydantic[email]==2.5.0
httpx==0.25.2
Pillow==10.1.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
fastapi-users==12.1.3
fastapi-users-db-sqlalchemy==6.0.1
websockets==12.0
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
jinja2==3.1.2